+++ apps/client/src/environments/patch-notes.ts
@@ -1,3 +1,16 @@
 export const patchNotes = `### Bug Fixes
 
-* **list:** fixed crafter icons not loading properly.`;
+* **alarms:** fixed custom sound input in timer settings popup.
+* **allagan-reports:** fixed fishing reports not being editable.
+* **desktop:** fixed character search.
+* **desktop:** fixed Universalis reporter.
+* **gearset:** fixed materias popup not adding materias properly.
+* **settings:** fixed sound notification setting not being applied properly.
+* **simulator:** added missing medicines.
+
+
+### Features
+
+* **db:** added link from container items to possible loots.
+* **db:** added link from npc pages to their shop content.
+* **db:** new allagan reports import.`;